refactor(apps/web/components): renamed ClothingClass to Category
