+++ apps/web/app/wardrobe/clothes/item/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { ClothingItem } from "@/components/wardrobe/clothing-item";
+import { ClothingItemEditor } from "@/components/wardrobe/clothing-item-editor";
 import { prisma } from "@/lib/prisma";
 import { getCategoryById, getColourById, getSizeById } from "@/lib/utils";
 import { notFound } from "next/navigation";
@@ -22,15 +22,8 @@ export default async function ItemPage({ params }: { params: Promise<{ id: strin
   const isWaterproof = !!item.waterproof;
 
   return (
-    <ClothingItem
-      src={item.item_url}
-      alt={item.item_name}
-      name={item.item_name}
-      category={category}
-      size={size}
-      colour={colour}
-      isWaterproof={isWaterproof}
-      environment={item.environment ?? undefined}
+    <ClothingItemEditor
+      item={item}
     />
   );
 }
