+++ apps/web/lib/utils.ts
@@ -1,4 +1,4 @@
-import { clothingClass, ClothingClass, slots } from "@/components/shared/clothing-class-select";
+import { categories, Category, slots } from "@/components/shared/category-select";
 import { Colour, colours } from "@/components/shared/colour-select";
 import { Size, sizes } from "@/components/shared/size-select";
 import { clsx, type ClassValue } from "clsx";
@@ -12,8 +12,8 @@ export function getColourId(colour: Colour): number {
   return colours.findIndex((item) => item.name === colour.name);
 }
 
-export function getCategoryId(c: ClothingClass): number {
-  return clothingClass.findIndex((item) => item === c);
+export function getCategoryId(c: Category): number {
+  return categories.findIndex((item) => item === c);
 }
 
 export function getSizeId(size: Size) {
@@ -24,14 +24,14 @@ export function getColourById(id: number): Colour | null {
   return colours[id] ?? null;
 }
 
-export function getCategoryById(id: number): ClothingClass | null {
-  return clothingClass[id] ?? null;
+export function getCategoryById(id: number): Category | null {
+  return categories[id] ?? null;
 }
 
 export function getSizeById(id: number): Size | null {
   return sizes[id] ?? null;
 }
 
-export function getSlot(c: ClothingClass): Number {
+export function getSlot(c: Category): Number {
   return slots.findIndex((item) => item.includes(c));
 }
